% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{bingo}
\alias{bingo}
\title{Create unique bingo cards with specified dimensions and custom base}
\usage{
bingo(n, dim = c(2, 3), base = 1:90, sort = TRUE, seed = NULL)
}
\arguments{
\item{n}{number of cards. Required.}

\item{dim}{card dimensions. Default is c(2,3)}

\item{base}{Base sequence. Default is 1:90. Can be a vector of numbers,
characters, character strings or image/icon file paths.}

\item{sort}{Flag if card sequence should be sorted. Default is TRUE.}

\item{seed}{seed}
}
\value{
list of banko_list class with elements of banko class
}
\description{
In bingo unique cards are created with each item occuring several times, but
only once on each board.
}
\examples{
bingo(2, base = 1:15, sort = TRUE)
bingo(2, base = list.files("/Users/au301842/Library/CloudStorage/OneDrive-Personal/Research/PhD/Formidling/Pictograms", full.names = TRUE))
}
