% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{cards}
\alias{cards}
\title{Creates n unique cards}
\usage{
cards(n, seed = NULL)
}
\arguments{
\item{n}{number of cards}

\item{seed}{integer seed. Default is NULL. The used seed is saved as attribute.}
}
\value{
list
}
\description{
Creates n unique cards
}
\examples{
cards(5) |>
  purrr::map(\(.x) attr(.x, which = "banko_seed")) |>
  unique()
attr(cards(5), which = "banko_seed")
}
