% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{cards_grob}
\alias{cards_grob}
\title{Multipage multicard PDF export}
\usage{
cards_grob(data, n = 5, note = "agdamsbo/banko")
}
\arguments{
\item{data}{list of gg_card}

\item{n}{number of cards per page}

\item{note}{note to print on}
}
\value{
list
}
\description{
Multipage multicard PDF export
}
\examples{
# Not evaluated
cards(5) |>
  purrr::map(gg_card) |>
  cards_grob() |>
  export_pdf(path = "banko.pdf")

bingo(5,
  base = list.files("/Users/au301842/Library/CloudStorage/OneDrive-Personal/Research/PhD/Formidling/Pictograms",
    full.names = TRUE
  )
) |>
  purrr::map(gg_card) |>
  cards_grob() |>
  export_pdf(path = "bingo.pdf")

memory(
  n = 10,
  base = list.files("/Users/au301842/Library/CloudStorage/OneDrive-Personal/Research/PhD/Formidling/Pictograms",
    full.names = TRUE
  )
) |>
  purrr::map(gg_card) |>
  cards_grob() |>
  export_pdf(path = "memory.pdf")

memory(
  n = 10,
  base = list.files("/Users/au301842/Downloads/twemoji-color-font-main/assets/twemoji-svg",
    full.names = TRUE
  )
) |>
  purrr::map(gg_card) |>
  cards_grob() |>
  export_pdf(path = "memory_emoji.pdf")
}
