% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{extend_seq}
\alias{extend_seq}
\title{Extend sequence with specified fill to match target length}
\usage{
extend_seq(data, target, fill = NA)
}
\arguments{
\item{data}{vector}

\item{target}{numeric}

\item{fill}{vector}
}
\value{
vector
}
\description{
Extend sequence with specified fill to match target length
}
\examples{
extend_seq(1:3, 8)
}
