% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{generate_bingo}
\alias{generate_bingo}
\title{Generate bingo boards}
\usage{
generate_bingo(dim = c(2, 3), base = 1:90, sort = FALSE)
}
\arguments{
\item{dim}{card dimensions. Default is c(2,3)}

\item{base}{Base sequence. Default is 1:90. Can be a vector of numbers,
characters, character strings or image/icon file paths.}

\item{sort}{Flag if card sequence should be sorted. Default is TRUE.}
}
\value{
matrix of banko class
}
\description{
Generate bingo boards
}
\examples{
generate_bingo()
}
