% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{get_sequence}
\alias{get_sequence}
\title{Convert df/tibble to vector. Option to remove NAs}
\usage{
get_sequence(data, no_nas = TRUE)
}
\arguments{
\item{data}{data frame, tibble or vector.}

\item{no_nas}{remove NAs. Default is TRUE}
}
\value{
vector
}
\description{
Convert df/tibble to vector. Option to remove NAs
}
\examples{
data <- c(1,2,NA)
get_sequence(c(1,NA,2))
get_sequence(c(1,2,NA),FALSE)
}
