% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{gg_seq_heat}
\alias{gg_seq_heat}
\title{Plot sample of base sequence as binary heatmap}
\usage{
gg_seq_heat(
  data,
  text.size = 14,
  ncol = NULL,
  base.seq = 1:90,
  colors = c("white", "grey")
)
}
\arguments{
\item{data}{numeric vector}

\item{text.size}{text size. Default 14.}

\item{ncol}{number of columns to plot. Defaults to ceiling(sqrt(length(base.seq)))}

\item{base.seq}{numeric vector to plot. Default is 1:90}

\item{colors}{colors for heatmap. Defaults is c("white","grey").}
}
\value{
ggplot2 list object
}
\description{
Plot sample of base sequence as binary heatmap
}
\examples{
cards(5)[[1]] |>
  get_sequence() |>
  gg_seq_heat(colors = c("white", "grey"))
c() |> gg_seq_heat(colors = c("white", "grey"), base.seq = sample(1:100, 25))
c() |> gg_seq_heat(colors = c("white", "grey"), base.seq = LETTERS)
}
