% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{is_unique_card}
\alias{is_unique_card}
\title{Test if card is unique compared to list of cards}
\usage{
is_unique_card(p, l)
}
\arguments{
\item{p}{new card}

\item{l}{list of cards}
}
\value{
logical
}
\description{
Test if card is unique compared to list of cards
}
\examples{
is_unique_card(c(1, 2, NA), list(c(2, 4, 6), c(1, NA, 2)))
}
