% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{is_unique_plate}
\alias{is_unique_plate}
\title{Test if plate is unique compared to list of plates}
\usage{
is_unique_plate(p, l)
}
\arguments{
\item{p}{new plate}

\item{l}{list of plates}
}
\value{
logical
}
\description{
Test if plate is unique compared to list of plates
}
\examples{
is_unique_plate(c(1,2,NA),list(c(2,4,6),c(1,NA,2)))
}
