% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{memory}
\alias{memory}
\title{Create memory card}
\usage{
memory(n = NULL, dim = c(2, 3), base = 1:90, sort = FALSE, seed = NULL)
}
\arguments{
\item{n}{number of cards. Default is NULL and will create all possible cards}

\item{dim}{card dimensions. Default is c(2,3)}

\item{base}{Base sequence. Default is 1:90}

\item{sort}{Flag if card sequence should be sorted.}

\item{seed}{seed}
}
\value{
list of banko_list class with elements of banko class
}
\description{
Each element only occurs once
}
\examples{
memory(n = 2)
memory(base = list.files("/Users/au301842/Library/CloudStorage/OneDrive-Personal/Research/PhD/Formidling/Pictograms", full.names = TRUE))
}
