% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{n_complete_rows}
\alias{n_complete_rows}
\title{Number of complete rows in each card from sequence}
\usage{
n_complete_rows(cards, sequence = NULL)
}
\arguments{
\item{cards}{banko cards}

\item{sequence}{number sequence}
}
\value{
numeric vector
}
\description{
Number of complete rows in each card from sequence
}
\examples{
cards(10) |>
  sequence4one() |>
  (\(.x) n_complete_rows(.x[[1]], .x[[2]]))()

n_complete_rows(cards = cards(40)) |>
  factor() |>
  summary()
}
