% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{n_each_card}
\alias{n_each_card}
\title{Gives number of correct fields on each card from sequence}
\usage{
n_each_card(cards, sequence = NULL)
}
\arguments{
\item{cards}{banko cards}

\item{sequence}{number sequence}
}
\value{
numeric vector
}
\description{
Gives number of correct fields on each card from sequence
}
\examples{
cards <- cards(5)
n_each_card(cards(3), sample(1:90, 30))
}
