% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{plates}
\alias{plates}
\title{Creates n random plates. No checking of uniqueness}
\usage{
plates(n)
}
\arguments{
\item{n}{number of plates}
}
\value{
list
}
\description{
Creates n random plates. No checking of uniqueness
}
\examples{
plates(5)
}
