% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{prepare_gg_card}
\alias{prepare_gg_card}
\title{Prepare matrix for gg_card}
\usage{
prepare_gg_card(data)
}
\arguments{
\item{data}{any object with two dimensions (data.frame or matrix)}
}
\value{
tibble
}
\description{
Prepare matrix for gg_card
}
\examples{
matrix(1:10, ncol = 5) |> prepare_gg_card()
matrix(seq(2 * 3), ncol = 2) |> prepare_gg_card()
}
