% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{seq_iter}
\alias{seq_iter}
\title{Iterative sequence}
\usage{
seq_iter(i, n)
}
\arguments{
\item{i}{iteration}

\item{n}{sequence size}
}
\value{
numeric vector
}
\description{
Iterative sequence
}
\examples{
for (i in 1:2) print(seq_iter(i, 5))
}
