% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{sequence4one}
\alias{sequence4one}
\title{Extracts shortest sequence of numbers to get at least one complete row on
each card}
\usage{
sequence4one(data, g = 100, selection = "min")
}
\arguments{
\item{data}{list of cards}

\item{g}{number of subsets to test}

\item{selection}{selection strategy for sequence. Can be "min" or "random".}
}
\value{
list with list and numeric vector
}
\description{
This can be used to draw a sequence before the game to ensure all will have
at least one row.

Could be used for travebanko with FDF, the scouts or others.
}
\examples{
cards(20) |>
  sequence4one(selection = "min") |>
  purrr::pluck("sequence") |>
  length()
cards(20) |>
  sequence4one(selection = "random") |>
  purrr::pluck("sequence") |>
  length()
}
