% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{shuffle_seq}
\alias{shuffle_seq}
\title{Randomly shuffles the order of a vector}
\usage{
shuffle_seq(data)
}
\arguments{
\item{data}{vector}
}
\value{
vector
}
\description{
Randomly shuffles the order of a vector
}
\examples{
1:10 |> shuffle_seq()
}
