% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{stats_walk}
\alias{stats_walk}
\title{Travebanko stats}
\usage{
stats_walk(cards, sequence, stops)
}
\arguments{
\item{cards}{banko cards list.}

\item{sequence}{numeric vector of sequence drawn.}

\item{stops}{n stops}
}
\value{
list
}
\description{
Travebanko stats
}
\examples{
cards(10) |> sequence4one() |>
 (\(.x){
   stats_walk(.x[[1]], .x[[2]], stops = 2)
 })()
}
