% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{stops_walk}
\alias{stops_walk}
\title{Travebanko stops}
\usage{
stops_walk(sequence, stops, header = "Post {sign.index}")
}
\arguments{
\item{sequence}{numeric vector of sequence drawn.}

\item{stops}{n stops}

\item{header}{header glue string}
}
\value{
list
}
\description{
Travebanko stops
}
\examples{
cards(10) |>
  sequence4one() |>
  (\(.x){
    stops_walk(.x[[2]], stops = 2)
  })()
}
