% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{travebanko}
\alias{travebanko}
\title{Travebanko grob merge}
\usage{
travebanko(
  data,
  stops,
  post.footer =
    "Post {sign.index} of {stops}. Put up on {format(Sys.Date(),'\%d-\%m-\%Y')}.",
  post.header = "Post {sign.index}"
)
}
\arguments{
\item{data}{banko cards}

\item{stops}{stops to make posters for}

\item{post.footer}{note for footer. Available values for the string are:
sign.index (current sign/stop), stops (number of stops) and the supplied
cards list (data).}

\item{post.header}{header glue string for signs/posts. Available values for
the string are: sign.index (current sign/stop) and stops (number of stops).}
}
\value{
list
}
\description{
Travebanko grob merge
}
\examples{
# data <- cards(30, 5)
# l <- cards(30, 5) |> travebanko(stops = 8)
# l |> export_pdf()
}
