% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cards.R
\name{unique_numbers}
\alias{unique_numbers}
\title{Extract unique numbers from cards and mix}
\usage{
unique_numbers(cards)
}
\arguments{
\item{cards}{banko cards}
}
\value{
vector
}
\description{
Extract unique numbers from cards and mix
}
\examples{
cards(20) |> unique_numbers()
}
