% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plates.R
\name{eliminate}
\alias{eliminate}
\title{Eliminate elements in columns to fullfill criteria for plates}
\usage{
eliminate(cols)
}
\arguments{
\item{cols}{tibble of complete columns}
}
\value{
tibble
}
\description{
Eliminate elements in columns to fullfill criteria for plates
}
