% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_pdf}
\alias{export_pdf}
\title{Print list of grobs as PDF}
\usage{
export_pdf(
  list,
  path = "banko_{attr(list, which = 'banko_seed')}.pdf",
  paper = "a4"
)
}
\arguments{
\item{list}{list of grobs to print}

\item{path}{output path. glue string}

\item{paper}{paper format. Passed to "pdf" engine.}
}
\description{
Print list of grobs as PDF
}
