% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{gg_card}
\alias{gg_card}
\title{Print banko cards with ggplot2}
\usage{
gg_card(data, text.size = 14, title = NULL, note = NULL, id.hash = FALSE)
}
\arguments{
\item{data}{"banko" class object}

\item{text.size}{size for number. Default is 14.}

\item{title}{optional title for all cards}

\item{note}{optional note for all cards}

\item{id.hash}{add md5 hash to notes section}
}
\value{
ggplot2 list object
}
\description{
Print banko cards with ggplot2
}
\examples{
data <- cards(5)[[1]]
data |> gg_card()
cards(5) |>
  purrr::map(gg_card)
}
