% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{plot.banko}
\alias{plot.banko}
\title{Base R style plotting}
\usage{
\method{plot}{banko}(x, ...)
}
\arguments{
\item{x}{"banko" class object}

\item{...}{not used. Conforms to standard docs for plot extensions}
}
\value{
plot
}
\description{
Base R style plotting
}
\examples{
cards(5) |> purrr::map(plot)

cards(5)[[1]] |> plot()
}
