% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{travebanko}
\alias{travebanko}
\title{Travebanko grob merge}
\usage{
travebanko(
  data,
  stops,
  post.footer = "Please make a note when they were put up and taken down, if in public."
)
}
\arguments{
\item{data}{banko cards}

\item{stops}{stops to make posters for}

\item{post.footer}{note for footer}
}
\value{
list
}
\description{
Travebanko grob merge
}
\examples{
l <- cards(30, 5) |> travebanko(stops = 8)
l |> export_pdf()
}
